refactor: split normalize file utils out into content and date utils
