+++ src/utils/dates.ts
@@ -3,47 +3,7 @@ import { compareDesc, isValid, parseISO } from 'date-fns'
 export type DateLike = Date | string | number | null | undefined
 
 /**
- * Remove known data file extensions from a content id
- *
- * - `example.yaml` -> `example`
- * - `parent/example.yaml` -> `parent/example`
- *
- * @param id - The content id to strip
- * @returns The content id without the `.yaml` extension
- */
-export function stripDataExtension(id: string): string {
-  return id.replace(/\.(ya?ml)$/i, '')
-}
-
-/**
- * Split a content id into section and nested itemId
- *
- * - `example.yaml` -> `example`
- * - `parent/example.yaml` -> `parent`, `example`
- *
- * @param id - The content id to parse
- * @returns The section, itemId, and whether it's a top-level entry
- */
-export function parseContentId(id: string): {
-  isTopLevel: boolean
-  itemId: string
-  section: string
-} {
-  const base = stripDataExtension(id)
-  const parts = base.split('/')
-  const section = parts[0] || ''
-  const itemId = parts.slice(1).join('/')
-
-  return {
-    isTopLevel: parts.length === 1,
-    itemId,
-    section,
-  }
-}
-
-/**
  * Check if an object has date fields
- *
  * @param x - The object to check
  * @returns True if the object has date fields
  */
@@ -53,7 +13,6 @@ function hasDateFields(x: unknown): x is { date?: DateLike; startDate?: DateLike
 
 /**
  * Get the primary date key (date or startDate) as a string
- *
  * @param input - The object to check
  * @returns The primary date key
  */
@@ -67,13 +26,11 @@ export function getDateKey(input: unknown): string {
       : ''
     )
   }
-
   return ''
 }
 
 /**
  * Convert a date-like value to a Date object
- *
  * @param value - The date-like value to convert
  * @returns The Date object, or null if the value is invalid
  */
@@ -101,12 +58,11 @@ export function toDate(value: DateLike): Date | null {
 
 /**
  * Get a value by path from an object
- *
  * @param obj - The object to get the value from
  * @param path - The path to the value
  * @returns The value, or undefined if the path is invalid
  */
-export function getByPath(obj: unknown, path: string): unknown {
+function getByPath(obj: unknown, path: string): unknown {
   if (obj === null || typeof obj !== 'object') {
     return undefined
   }
@@ -125,21 +81,24 @@ export function getByPath(obj: unknown, path: string): unknown {
   return current
 }
 
-export function firstDefinedByPaths(obj: unknown, paths: string[]): unknown {
+/**
+ * Get the first defined value from multiple paths
+ * @param obj - The object to check
+ * @param paths - The paths to check
+ * @returns The first defined value, or undefined if none found
+ */
+function firstDefinedByPaths(obj: unknown, paths: string[]): unknown {
   for (const p of paths) {
     const v = getByPath(obj, p)
-
     if (v !== undefined) {
       return v
     }
   }
-
   return undefined
 }
 
 /**
- * Sort items by date/startDate descending
- *
+ * Sort items by date/startDate descending using an accessor function
  * @param items - The items to sort
  * @param accessor - The accessor function to get the date
  * @returns The sorted items
@@ -163,12 +122,11 @@ export function sortByAccessorDateDesc<T>(items: T[], accessor: (item: T) => Dat
 
 /**
  * Sort items by date keys descending
- *
  * @param items - The items to sort
  * @param dateKeys - The date keys to sort by
  * @returns The sorted items
  */
-export function sortByDateKeysDesc<T>(items: T[], dateKeys: string[]) {
+function sortByDateKeysDesc<T>(items: T[], dateKeys: string[]) {
   return sortByAccessorDateDesc(items, (item) => {
     const v = firstDefinedByPaths(item, dateKeys)
     return v instanceof Date || typeof v === 'string' || typeof v === 'number' ? v : undefined
@@ -177,7 +135,6 @@ export function sortByDateKeysDesc<T>(items: T[], dateKeys: string[]) {
 
 /**
  * Sort items by date/startDate descending
- *
  * @param items - The items to sort
  * @param getValue - The accessor function to get the date
  * @returns The sorted items
